blinky-server/pyproject.toml
[project]
name = "blinky-server"
version = "0.1.0"
description = "Fleet management server for Blinky Time LED art devices"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "pyserial>=3.5",
    "pyserial-asyncio-fast>=0.13",
    "pydantic>=2.0",
    "websockets>=13.0",
]

[project.optional-dependencies]
ble = ["bleak>=0.22"]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "httpx>=0.27",           # FastAPI TestClient async support
    "mypy>=1.11",
    "ruff>=0.7",
]

[project.scripts]
blinky-server = "blinky_server.__main__:main"

# ── mypy ──
[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true

[[tool.mypy.overrides]]
module = [
    "serial.*",
    "serial_asyncio_fast.*",
    "bleak.*",
    "uvicorn.*",
]
ignore_missing_imports = true

# ── ruff ──
[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
    "TCH",  # flake8-type-checking
    "RUF",  # ruff-specific
]
ignore = [
    "E501",   # line length (handled by formatter)
    "B008",   # function call in default arg (FastAPI Depends pattern)
    "TC001",  # type-checking imports - conflicts with runtime usage in fixtures
    "TC002",  # type-checking imports for third-party
    "TC003",  # type-checking imports for stdlib
]

[tool.ruff.lint.isort]
known-first-party = ["blinky_server"]

# ── pytest ──
[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
filterwarnings = [
    "ignore::DeprecationWarning",
]
